+++ src/ua/data_types/data_change_filter.rs
@@ -1,0 +1,9 @@
+use crate::{ua, MonitoringFilter};
+
+crate::data_type!(DataChangeFilter);
+
+impl MonitoringFilter for DataChangeFilter {
+    fn to_extension_object(&self) -> ua::ExtensionObject {
+        ua::ExtensionObject::new(self)
+    }
+}