+++ src/ua/data_types/element_operand.rs
@@ -1,0 +1,9 @@
+use crate::{ua, FilterOperand};
+
+crate::data_type!(ElementOperand);
+
+impl FilterOperand for ElementOperand {
+    fn to_extension_object(&self) -> ua::ExtensionObject {
+        ua::ExtensionObject::new(self)
+    }
+}